+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix ``datafields`` for ``ExtensionFields``
+  [gotcha]
 
 
 1.0.1 (2018-06-11)